Added more options to use custom builds for tests
To install custom builds we can now use:
- custom_build_url
- custom_build_repo_url
- custom_build_path

Generally the community developers would make changes in their code
and fire make glusterrpms, so they would be having the rpms but
not the url, which blocks them to use this utility to check the
performance of the changes made in their sandbox. With this option
provided it will become very easy for any body to check their changes
by using custom_build_path option.

Signed-off-by: Rinku Kothiya <rkothiya@redhat.com>